Implement initializeLives(). Add more Javadocs. Display gameoverBanner when all lives lost.
Ball now resets it's self when hitting bottom of border. When ball hits bottom of border, a life is removed from the screen. When all three lives are removed from the screen, the game ends and the GAMEOVER banner is displayed. Still need to fix the dimensions and positioning of the GAMEOVER banner.